mapeo de base de datos

+++ Cruz_FrancoMatias_2931_TP4/src/main/java/ar/edu/unju/fi/tp4/service/IProductoService.java
@@ -16,7 +16,7 @@ public interface IProductoService {
 	
 	public List<Producto> getAllProductos();
 	
-	public Producto getProductoporCodigo(int codigo);
+	public Producto getProductoporCodigo(long codigo);
